+++ src/readoutErrorCodes.h
@@ -51,5 +51,6 @@
   { 3241, "Inconsistent data", nullptr},
   { 3242, "Database error", nullptr},
   { 3243, "Control problem", nullptr},
+  { 3244, "Should not happen problem", nullptr},
 */